requirements.txt
pygame-ce
python-dotenv
groq
